+++ src/Serilog.Sinks.Http/Sinks/Http/Private/IO/DirectoryService.cs
@@ -1,33 +1,25 @@
-﻿// Copyright 2015-2019 Serilog Contributors
-//
+﻿// Copyright 2015-2018 Serilog Contributors
+// 
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
-//
+// 
 //     http://www.apache.org/licenses/LICENSE-2.0
-//
+// 
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System.Net.Http;
-using System.Threading.Tasks;
+using System.IO;
 
-namespace Serilog.Sinks.Http.Private.Network
+namespace Serilog.Sinks.Http.Private.IO
 {
-    public class DefaultHttpClient : IHttpClient
+    public class DirectoryService : IDirectoryService
     {
-        private readonly HttpClient client;
-
-        public DefaultHttpClient() =>
-            client = new HttpClient();
-
-        public Task<HttpResponseMessage> PostAsync(string requestUri, HttpContent content) =>
-            client.PostAsync(requestUri, content);
-
-        public void Dispose() =>
-            client.Dispose();
+        public string[] GetFiles(string path, string searchPattern) =>
+            Directory.GetFiles(path, searchPattern);
     }
 }
+