+++ src/Serilog.Sinks.Http/Sinks/Http/Private/Network/PayloadReader.cs
@@ -11,7 +11,7 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-#pragma warning disable CS1591 // Missing XML comment for publicly visible type or member
+
 using System;
 using System.Collections.Generic;
 using System.IO;
@@ -117,4 +117,3 @@ namespace Serilog.Sinks.Http.Private.Network
         }
     }
 }
-#pragma warning restore CS1591 // Missing XML comment for publicly visible type or member