+++ test/Serilog.Sinks.HttpTests/DurableHttpSinkUsingFileSizeRolledBuffersGivenCodeConfigurationShould.cs
@@ -6,18 +6,16 @@ using Serilog.Support;
 
 namespace Serilog
 {
-    public class DurableHttpSinkGivenCodeConfigurationShould : SinkFixture
+    public class DurableHttpSinkUsingFileSizeRolledBuffersGivenCodeConfigurationShould : SinkFixture
     {
-        public DurableHttpSinkGivenCodeConfigurationShould()
+        public DurableHttpSinkUsingFileSizeRolledBuffersGivenCodeConfigurationShould()
         {
             DeleteBufferFiles();
-
-#pragma warning disable CS0618 // Type or member is obsolete
 
             Logger = new LoggerConfiguration()
                 .MinimumLevel.Verbose()
                 .WriteTo
-                .DurableHttp(
+                .DurableHttpUsingFileSizeRolledBuffers(
                     requestUri: "some/route",
                     batchPostingLimit: 100,
                     period: TimeSpan.FromMilliseconds(1),
@@ -25,8 +23,6 @@ namespace Serilog
                     batchFormatter: new DefaultBatchFormatter(),
                     httpClient: new HttpClientMock())
                 .CreateLogger();
-
-#pragma warning restore CS0618 // Type or member is obsolete
         }
 
         protected override Logger Logger { get; }